% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rand.R
\name{create_rand}
\alias{create_rand}
\alias{get_rand}
\title{Random data set creation}
\usage{
create_rand(filename = NULL, n = 10000, seed = NULL)

get_rand()
}
\arguments{
\item{filename}{A filename (should have a .csv extension)}

\item{n}{Number of rows}

\item{seed}{Optional seed}
}
\description{
A simple function for creating a n by 2 matrix filled with random numbers. The
matrix is stored as a csv file. The function \code{get_rand} returns where the file
is stored.
}
\examples{
\dontrun{
create_rand("inst/raw-data/rand.csv", n=1e5, 1)
}
}
